Added tests for problems 1 - 12

+++ c/common/include/utils.h
@@ -7,6 +7,6 @@ void * x_malloc (size_t);
 void * x_calloc (size_t, size_t);
 void * x_realloc (void *, size_t);
 
-void swap_chars (char *, int, int);
+void swap_chars (char *, size_t, size_t);
 
 #endif
